Added UI and clicker art assets
Added UI elements to be used in the clicker: an icon, an animation (blood splatter) for feedback, buttons, more icons for upgrades, etc. Added a bg for the clicker game. Added taskbar sprites (and reused the asset for browser tabs). Changed positions of a few things (graphic/organizational changes only).