Finish up the changes for tonight (see below)
I added and finished the Audio Mixer. You can now change the sound levels using the gear icon in taskbar in Desktop, and the game saves the values. It's still a bit buggy and might not work perfectly all the time, but I'll work on the bugs later. Just let me know when you encounter one.